refactor: simplify conv chain & update package structure
