--- api/chatbot/chains/__init__.py
@@ -1,3 +1,0 @@
-from chatbot.chains.conversation import ConversationChain
-
-__all__ = ["ConversationChain"]
